[20] Move logic to s/subtree.go
